This patch syncs the gpu dev branch with trunk.  It also fixes a handful of bugs to get MPI working again.  This patch is full of debugger print messages, I'm just committing it to get the bug fixes in.  I'll be removing those messages shortly

+++ src/CCA/Components/MPM/ConstitutiveModel/MPMMaterial.cc
@@ -254,7 +254,7 @@ ScalarDiffusionModel* MPMMaterial::getScalarDiffusionModel() const
 }
 
 particleIndex MPMMaterial::createParticles(
-                                  CCVariable<short int>& cellNAPID,
+                                  CCVariable<int>& cellNAPID,
                                   const Patch* patch,
                                   DataWarehouse* new_dw)
 {
